+++ test/validator/samples/a11y-anchor-has-content/input.html
@@ -1,0 +1,1 @@
+<a href='/foo'></a>
