+++ test/validator/samples/a11y-anchor-has-content/warnings.json
@@ -1,0 +1,8 @@
+[{
+	"message": "A11y: <a> element should have child content",
+	"loc": {
+		"line": 1,
+		"column": 0
+	},
+	"pos": 0
+}]
